mnist_project/requirements.txt
torch
torchvision
streamlit
streamlit-drawable-canvas
psycopg2
Pillow
python-dotenv
pandas